fix: fix pagination main view
