+++ src/components/main-view/main-view.js
@@ -1,89 +1,15 @@
 import React, { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useTranslation } from 'react-i18next';
 import ArticleList from '../article-list/article-list';
-import { Tab, TabsList } from './styled-main-view';
+import { TabsList } from './styled-main-view';
 import {
   setTab,
   setTag,
   requestArticleAllPage,
   requestArticleFeed,
-  requestArticleByTag,
 } from '../../store/articleSlice';
 import { LOCAL_STORE_TOKEN_NAME } from '../../constants/consts';
-
-const YourFeedTab = (props) => {
-  const {
-    tab,
-    setPage,
-    page,
-    token,
-  } = props;
-  const { t } = useTranslation();
-  const dispatch = useDispatch();
-  const clickHandler = (ev) => {
-    ev.preventDefault();
-    setPage(0);
-    dispatch(setTab('feed'));
-    dispatch(setTag(''));
-    dispatch(requestArticleFeed());
-  };
-  useEffect(() => {
-    dispatch(requestArticleFeed(page));
-  }, [dispatch, page]);
-
-  if (!token) return null;
-
-  return (
-    <li>
-      <Tab href="" active={tab === 'feed'} onClick={clickHandler}>
-        {t('yourFeed')}
-      </Tab>
-    </li>
-  );
-};
-
-const GlobalFeedTab = ({ tab, setPage, page }) => {
-  const { t } = useTranslation();
-  const dispatch = useDispatch();
-  const clickHandler = (ev) => {
-    ev.preventDefault();
-    setPage(0);
-    dispatch(setTab('all'));
-    dispatch(setTag(''));
-    dispatch(requestArticleAllPage());
-  };
-  useEffect(() => {
-    dispatch(requestArticleAllPage(page));
-  }, [dispatch, page]);
-
-  return (
-    <li>
-      <Tab href="" active={tab === 'all'} onClick={clickHandler}>
-        {t('feed')}
-      </Tab>
-    </li>
-  );
-};
-
-const TagFilterTab = ({ tag, setPage, page }) => {
-  const dispatch = useDispatch();
-  if (!tag) return null;
-  if (tag === '') setPage(0);
-  dispatch(setTab(''));
-  useEffect(() => {
-    dispatch(requestArticleByTag({ tag, page }));
-  }, [dispatch, tag, page]);
-
-  return (
-    <li>
-      <Tab to="" active>
-        #
-        {tag}
-      </Tab>
-    </li>
-  );
-};
+import { GlobalFeedTab, TagFilterTab, YourFeedTab } from './tabs';
 
 const MainView = () => {
   const dispatch = useDispatch();
