+++ src/components/main-view/tabs/index.js
@@ -1,0 +1,3 @@
+export { default as TagFilterTab } from './tag-filter-tab/tag-filter-tab';
+export { default as GlobalFeedTab } from './global-feed-tab/global-feed-tab';
+export { default as YourFeedTab } from './your-feed-tab/your-feed-tab';
